Args and return for callback start_requests/1
This callback is called when the spider is opened for scraping.

It is called when no `start_urls` provided.

If `init/1` callback returns an integer as the third element in return
tuple, the `start_requests` will be delayed that milliseconds before
being called.

`start_requests` must return `{ok, list(), state()}`